% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProQuestSource.R
\name{ProQuestSource}
\alias{ProQuestSource}
\title{A Source to handle .eml output from ProQuest}
\usage{
ProQuestSource(x)
}
\arguments{
\item{x}{A filename string or connection}
}
\value{
An object of class \code{ProQuestSource}, inheriting from
\link[tm:SimpleSource]{tm::SimpleSource}.
}
\description{
Constructs a Source to handle documents sent by email from
ProQuest, in .eml format.
}
\details{
An attempt is made to fill up the standard \code{tm} metadata fields if
explicit mappings are not provided (or produce NULL). The \code{datetimestamp}
output field is taken as the first available of the following:
\code{datetimestamp} input field, \code{modtime} input field, \code{firstpubtime} input
field, \code{fetchtime} input field, or (if all else fails) the current date and
time, with a warning. \code{author} is taken by concatenating (with commas) the
\code{bylines} input field. \code{description} is taken from the input \code{summary}
field, and \code{heading} from \code{headline}. \code{origin} is taken from \code{source},
failing which the filename of the input file is used.

The returned documents have IDs constructed to be unique under all normal
circumstances. This is based on the input file's \code{source} field (if
available), the document date (drawn from the input file's \code{modtime},
failing that \code{firstpubtime}, failing that \code{fetchtime}, failing that the
current date with a warning), the document's sequence number in the input
file, and an 8 character MD5 digest of the document's metadata list.
}
\examples{
\dontrun{

s <- ProQuestSource('input.eml')
corp <- VCorpus(s)
}
}
