% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProQuest.R
\name{readProQuest}
\alias{readProQuest}
\title{A Reader function generator for ProQuest .eml data}
\usage{
readProQuest()
}
\value{
A \link[tm:Reader]{tm::Reader} function which returns objects inheriting from
\link[tm:PlainTextDocument]{tm::PlainTextDocument}
}
\description{
A Reader function generator for ProQuest .eml data
}
\seealso{
\link{ProQuestSource}
}
