% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseLanguage.R
\name{standardiseLanguage}
\alias{standardiseLanguage}
\title{Take the language field given by ProQuest and try to convert into a
two-letter ISO_639_2 code.}
\usage{
standardiseLanguage(v, tid)
}
\arguments{
\item{v}{A character vector containing the language string to look up
(only the first element is used).}

\item{tid}{An ID, for use in error reporting}
}
\value{
A string representing a ISO 639:2 language code, or character(0)
}
\description{
Constructs a Source to handle documents sent by email from
ProQuest, in .eml format.
}
