% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm.plugin.proquest-package.R
\docType{package}
\name{tm.plugin.proquest-package}
\alias{tm.plugin.proquest}
\alias{tm.plugin.proquest-package}
\title{tm.plugin.proquest: Import Articles from ProQuest Using the 'tm' Text Mining Framework}
\description{
Provides a tm Source function to create corpora from articles
  emailed from the ProQuest database.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/pmyteh/tm.plugin.proquest}
  \item Report bugs at \url{http://github.com/pmyteh/tm.plugin.proquest/issues}
}

}
\author{
\strong{Maintainer}: Tom Nicholls \email{T.E.Nicholls@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-6971-8614}{ORCID})

}
\keyword{internal}
